Made Kros.AspNetCore target solely .NET Core 3.0.
