+++ src/Kros.AspNetCore/BaseStartup.cs
@@ -4,13 +4,8 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-
-#if (NETCOREAPP3_0)
-
 using Microsoft.Extensions.Hosting;
-
-#endif
+using Microsoft.Extensions.Logging;
 
 namespace Kros.AspNetCore
 {
@@ -19,27 +14,6 @@ namespace Kros.AspNetCore
     /// </summary>
     public abstract class BaseStartup
     {
-#if (NETCOREAPP2_2)
-
-        /// <summary>
-        /// Ctor.
-        /// </summary>
-        /// <param name="env">Information about the web hosting environment.</param>
-        public BaseStartup(IHostingEnvironment env)
-        {
-            IConfigurationBuilder builder = new ConfigurationBuilder()
-                .SetBasePath(env.ContentRootPath)
-                .AddJsonFile("appsettings.json")
-                .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true)
-                .AddJsonFile($"appsettings.local.json", optional: true)
-                .AddEnvironmentVariables();
-
-            Configuration = builder.Build();
-            Environment = env;
-        }
-
-#else
-
         /// <summary>
         /// Ctor.
         /// </summary>
@@ -57,28 +31,15 @@ namespace Kros.AspNetCore
             Environment = env;
         }
 
-#endif
-
         /// <summary>
         /// Application configuration.
         /// </summary>
         public IConfiguration Configuration { get; }
 
-#if (NETCOREAPP2_2)
-
-        /// <summary>
-        /// Application environment.
-        /// </summary>
-        public IHostingEnvironment Environment { get; }
-
-#else
-
         /// <summary>
         /// Application environment.
         /// </summary>
         public IWebHostEnvironment Environment { get; }
-
-#endif
 
         /// <summary>
         /// Application services configuration.
