+++ src/Kros.AspNetCore/Environments/EnvironmentsExt.cs
@@ -1,32 +1,28 @@
-﻿#if (NETCOREAPP2_2)
-
-namespace Microsoft.AspNetCore.Hosting
+﻿namespace Microsoft.Extensions.Hosting
 {
     /// <summary>
     /// Valid environments.
     /// </summary>
-    public static class Environments
+    public static class EnvironmentsExtended
     {
         /// <summary>
         /// Development environment.
         /// </summary>
-        public const string Development = "Development";
+        public static readonly string Development = Environments.Development;
 
         /// <summary>
         /// Test environment.
         /// </summary>
-        public const string Test = "Test";
+        public static readonly string Test = "Test";
 
         /// <summary>
         /// Staging environment.
         /// </summary>
-        public const string Staging = "Staging";
+        public static readonly string Staging = Environments.Staging;
 
         /// <summary>
         /// Production environment.
         /// </summary>
-        public const string Production = "Production";
+        public static readonly string Production = Environments.Production;
     }
 }
-
-#endif
