+++ src/Kros.AspNetCore/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using Kros.AspNetCore;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using System;
 
@@ -10,15 +9,6 @@ namespace Microsoft.Extensions.DependencyInjection
     /// </summary>
     public static class ServiceCollectionExtensions
     {
-#if (NETCOREAPP2_2)
-
-        /// <summary>
-        /// Default MVC compatibility version for <see cref="AddWebApi(IServiceCollection)"/>.
-        /// </summary>
-        public static CompatibilityVersion DefaultCompatibilityVersion => CompatibilityVersion.Version_2_2;
-
-#endif
-
         /// <summary>
         /// Configure options of type <typeparamref name="TOptions"/> and binds it to section with the name
         /// same as the class <typeparamref name="TOptions"/> without Options suffix in <paramref name="configuration"/>.
@@ -47,48 +37,14 @@ namespace Microsoft.Extensions.DependencyInjection
             string sectionName) where TOptions : class
             => services.Configure<TOptions>(options => configuration.GetSection(sectionName).Bind(options));
 
-#if (NETCOREAPP2_2)
-
-        /// <summary>
-        /// Adds the minimum essential MVC services to the DI container for web API services.
-        /// (MVC Core, JSON Formatters, CORS, API Explorer, Authorization)
-        /// Additional services must be added separately using the <see cref="IMvcCoreBuilder"/> returned from this method.
-        /// As compatibility version for MVC is used <see cref="DefaultCompatibilityVersion"/>.
-        /// </summary>
-        /// <param name="services">MVC Core builder.</param>
-        public static IMvcCoreBuilder AddWebApi(this IServiceCollection services)
-            => AddWebApi(services, DefaultCompatibilityVersion);
-
-        /// <summary>
-        /// Adds the minimum essential MVC services to the DI container for web API services.
-        /// (MVC Core, JSON Formatters, CORS, API Explorer, Authorization)
-        /// Additional services must be added separately using the <see cref="IMvcCoreBuilder"/> returned from this method.
-        /// </summary>
-        /// <param name="services">MVC Core builder.</param>
-        /// <param name="compatibilityVersion">Compatibility version for MVC.</param>
-        public static IMvcCoreBuilder AddWebApi(this IServiceCollection services, CompatibilityVersion compatibilityVersion)
-        {
-            IMvcCoreBuilder builder = services.AddMvcCore().SetCompatibilityVersion(compatibilityVersion);
-
-            return builder.AddFormatterMappings()
-                .AddJsonFormatters()
-                .AddCors()
-                .AddApiExplorer()
-                .AddAuthorization();
-        }
-
-#else
-
         /// <summary>
         /// Adds the minimum essential MVC services to the DI container for web API services.
         /// (MVC Core, JSON Formatters, CORS, API Explorer, Authorization)
         /// Additional services must be added separately using the <see cref="IMvcBuilder"/> returned from this method.
         /// </summary>
         /// <param name="services">MVC builder.</param>
-        [Obsolete("AddWebApi is deprecated, please use AddControllers() instead.")]
+        [Obsolete("AddWebApi() is deprecated, please use AddControllers() instead.")]
         public static IMvcBuilder AddWebApi(this IServiceCollection services)
             => services.AddControllers();
-
-#endif
     }
 }
